+++ __tests__/company.test.ts
@@ -1,7 +1,10 @@
 import vnstock from "../src";
 import { Company } from "../src/core/stock/company";
 
-describe("Company", () => {
+// Skipped: VCI GraphQL endpoint (trading.vietcap.com.vn/data-mt/graphql)
+// returns HTTP 200 with empty body {} since ~2026-04. Planned migration to
+// KBS data source in v1.4. See: https://github.com/ttqteo/vnstock-js/issues
+describe.skip("Company", () => {
   let company: Company;
 
   beforeAll(() => {